TEST: Split test_statistics into two parts
First part works with Python >=  3.9

Second part works with Python >= 3.10
