+++ integration_tests/CMakeLists.txt
@@ -741,7 +741,8 @@ RUN(NAME generics_array_01   LABELS cpython llvm c)
 RUN(NAME generics_array_02   LABELS cpython llvm c)
 RUN(NAME generics_array_03   LABELS cpython llvm c)
 RUN(NAME generics_list_01    LABELS cpython llvm c)
-RUN(NAME test_statistics     LABELS cpython llvm NOFAST)
+RUN(NAME test_statistics_01  LABELS cpython llvm NOFAST)
+RUN(NAME test_statistics_02  LABELS cpython llvm NOFAST)
 RUN(NAME test_str_attributes LABELS cpython llvm c)
 RUN(NAME kwargs_01           LABELS cpython llvm c NOFAST)
